+++ src/main/java/ba/edu/ibu/finance_tracker/core/model/Expense.java
@@ -1,17 +1,18 @@
-package ba.edu.ibu.finance_tracker.rest.model;
+package ba.edu.ibu.finance_tracker.core.model;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 @Document
-public class RepeatingExpense {
+public class Expense {
 
     @Id
     private String id;
     private String userId;
     private double amount;
+    private String spentOn;
     private String category;
-    private String dueDate;
+    private String source;
 
     public String getId() {
         return id;
@@ -37,6 +38,14 @@ public class RepeatingExpense {
         this.amount = amount;
     }
 
+    public String getSpentOn() {
+        return spentOn;
+    }
+
+    public void setSpentOn(String spentOn) {
+        this.spentOn = spentOn;
+    }
+
     public String getCategory() {
         return category;
     }
@@ -45,12 +54,12 @@ public class RepeatingExpense {
         this.category = category;
     }
 
-    public String getDueDate() {
-        return dueDate;
+    public String getSource() {
+        return source;
     }
 
-    public void setDueDate(String dueDate) {
-        this.dueDate = dueDate;
+    public void setSource(String source) {
+        this.source = source;
     }
 
 }
