+++ src/main/java/ba/edu/ibu/finance_tracker/core/repository/AlertRepository.java
@@ -1,7 +1,8 @@
-package ba.edu.ibu.finance_tracker.rest.repository;
+package ba.edu.ibu.finance_tracker.core.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
-import ba.edu.ibu.finance_tracker.rest.model.Alert;
+
+import ba.edu.ibu.finance_tracker.core.model.Alert;
 
 public interface AlertRepository extends MongoRepository<Alert, String> {
