+++ src/main/java/ba/edu/ibu/finance_tracker/core/repository/ExpenseRepository.java
@@ -1,8 +1,9 @@
-package ba.edu.ibu.finance_tracker.rest.repository;
+package ba.edu.ibu.finance_tracker.core.repository;
 
 import java.util.List;
 import org.springframework.data.mongodb.repository.MongoRepository;
-import ba.edu.ibu.finance_tracker.rest.model.Expense;
+
+import ba.edu.ibu.finance_tracker.core.model.Expense;
 
 public interface ExpenseRepository extends MongoRepository<Expense, String> {
     List<Expense> findByUserId(String userId);
