+++ src/main/java/ba/edu/ibu/finance_tracker/core/repository/RepeatingExpenseRepository.java
@@ -1,9 +1,10 @@
-package ba.edu.ibu.finance_tracker.rest.repository;
+package ba.edu.ibu.finance_tracker.core.repository;
 
 import java.util.List;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
-import ba.edu.ibu.finance_tracker.rest.model.RepeatingExpense;
+
+import ba.edu.ibu.finance_tracker.core.model.RepeatingExpense;
 
 public interface RepeatingExpenseRepository extends MongoRepository<RepeatingExpense, String> {
     List<RepeatingExpense> findByUserId(String userId);
