+++ src/main/resources/application.yml
@@ -1,0 +1,10 @@
+
+configuration:
+
+  swagger:
+    app: 'FInance Management System'
+    description: 'A simple system built with Spring Boot'
+    developer: 'Emir Kugic'
+    email: 'emir.kugic@ibu.edu.ba'
+  mailsender:
+    default: 'sendgrid'
